Move database_status.py to a common place
`database_status.py` is used by the client and server part of the
CodeChecker too. To prepare source repository split up we move it to
a common place.
